+++ packages/blackbox_flutter/test/persistent_shared_prefs_store_test.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: deprecated_member_use
 import 'package:blackbox/blackbox.dart';
 import 'package:blackbox_flutter/blackbox_flutter.dart';
 import 'package:flutter_test/flutter_test.dart';